Move packages to match the organization of other modules, add the ability to chain filters for registry monitors
