--- library/core/registry/src/main/java/org/quiltmc/qsl/registry/event/api/RegistryEntryPredicate.java
@@ -1,8 +1,0 @@
-package org.quiltmc.qsl.registry.event.api;
-
-import net.minecraft.util.Identifier;
-
-@FunctionalInterface
-public interface RegistryEntryPredicate<V> {
-	boolean test(V entry, Identifier id, int rawId);
-}
