+++ library/core/registry/src/main/java/org/quiltmc/qsl/registry/impl/event/RegistryEventStorage.java
@@ -1,8 +1,8 @@
-package org.quiltmc.qsl.registry.event.impl;
+package org.quiltmc.qsl.registry.impl.event;
 
 import net.minecraft.util.registry.Registry;
 import org.quiltmc.qsl.base.api.event.ArrayEvent;
-import org.quiltmc.qsl.registry.event.api.RegistryEvents;
+import org.quiltmc.qsl.registry.api.event.RegistryEvents;
 
 public interface RegistryEventStorage<V> {
 	ArrayEvent<RegistryEvents.EntryAdded<V>> quilt$getEntryAddedEvent();
