+++ library/core/registry/src/main/java/org/quiltmc/qsl/registry/mixin/SimpleRegistryMixin.java
@@ -5,8 +5,8 @@ import net.minecraft.util.registry.Registry;
 import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.util.registry.SimpleRegistry;
 import org.quiltmc.qsl.base.api.event.ArrayEvent;
-import org.quiltmc.qsl.registry.event.api.RegistryEvents;
-import org.quiltmc.qsl.registry.event.impl.RegistryEventStorage;
+import org.quiltmc.qsl.registry.api.event.RegistryEvents;
+import org.quiltmc.qsl.registry.impl.event.RegistryEventStorage;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
